Last commit
This marks the end of development for this project.
It started as an ambitious overkill school project.
I have learned a great deal and am now ready to move on to other things.

This commit includes code cleanup, minor bug and syntax fixes, and an informative README.
Also added a Battle Royal mode as the last feture of the school project,
making it playable in its current state.

I hope that one day in the future, I will be able to recreate it as a
full-blown real game once I am more experienced.